==> Data Preparation/6- Downloading GMS eligibilty numbers.ipynb <==
from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.webdriver.chrome.options import Options

import time
import win32api
import pandas as pd
import os



# Function to check if the file has been downloaded
def wait_for_download(directory):
    # Wait for the download to finish (check every second)
    while True:
        files = os.listdir(directory)
        if files:  # Assuming there's only one file downloading at a time
            if not any(f.endswith('.crdownload') for f in files):  # Chrome temporary download extension
                return files[0]
        time.sleep(1)

# Function to rename the downloaded file
def rename_downloaded_file(download_dir, old_name, new_name):
    old_path = os.path.join(download_dir, old_name)
    new_path = os.path.join(download_dir, new_name)
    os.rename(old_path, new_path)



# Create a WebDriver instance (in this example, using Chrome)

download_dir = "C:\\Users\ \\ \"  # Change this to your preferred directory
os.makedirs(download_dir, exist_ok=True)

# Chrome options for headless download
options = webdriver.ChromeOptions()
options.add_argument('--headless')
options.add_argument('--disable-gpu')  # Sometimes needed for headless mode
options.add_argument(f'--no-sandbox')  # Bypass OS security model
options.add_argument('--disable-dev-shm-usage')  # Overcome limited resource problems
options.add_experimental_option("prefs", {
    "download.default_directory": download_dir,
    "download.prompt_for_download": False,  # Disable prompts
    "download.directory_upgrade": True,
    "safebrowsing.enabled": True  # Enable to prevent download error notifications
})

driver = webdriver.Chrome(options=options)


driver.get("https://www.sspcrs.ie/portal/annual-reporting/report/eligibility")
time.sleep(5)

reportPage = driver.find_element(By.XPATH, "/html/body/div/div/section/section[2]/section/div/div[5]/div[2]/h3/form/button")
reportPage.click()
time.sleep(10)

 monthList= ['201701','201702','201703','201704','201705','201706','201707','201708','201709','201710','201711','201712',
             '201801','201802','201803','201804','201805','201806','201807','201808','201809','201810','201811','201812',
             '201901','201902','201903','201904','201905','201906','201907','201908','201909','201910','201911','201912',
            '202001','202002','202003','202004','202005','202006','202007','202008','202009','202010','202011','202012',
            '202101','202102','202103','202104','202106','202107','202108','202109','202110','202111','202112',
             '202201','202202','202203','202204','202205','202206','202207','202208','202209','202210','202211','202212',
             '202301','202302','202303','202304','202305','202306','202307','202308','202309','202310','202311','202312',
            '202401','202402','202403','202404','202405','202406','202407','202408','202409','202410']


for month in monthList:
    monthCombo = driver.find_element(By.XPATH, "//img[contains(@id, 'dropdownIcon')]")
    monthCombo.click()
    
    wait = WebDriverWait(driver, 10)
    monthChoice = wait.until(EC.presence_of_element_located((By.XPATH, "//*[@title='"+month+"']")))
    monthChoice.click()
    time.sleep(5)
    applyButton = driver.find_element(By.XPATH, "//*[@id='gobtn']")
    applyButton.click()
    time.sleep(5)
    
    exportLink = driver.find_element(By.XPATH, "//a[@title='Export to different format']")
    exportLink.click()
    time.sleep(5)
    
    td_element = driver.find_element(By.XPATH, "//td[text()='Excel']")
    # Click on the <td> element
    td_element.click()
    time.sleep(5)
    td_element2 = driver.find_element(By.XPATH, "//td[contains(text(), 'Excel 2007')]")
    # Click on the <td> element
    td_element2.click()
    time.sleep(5)
    # After triggering the download
    #downloaded_file_name = wait_for_download(download_dir)  # Get the name of the downloaded file
    new_file_name = month+".xlsx"  # Set your desired new file name
    rename_downloaded_file(download_dir, "MC table breakdown_dash.xlsx", new_file_name)

    print(f"File downloaded and renamed to {new_file_name}")
    time.sleep(10)
    print("Download Done")
    
driver.close()

from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.webdriver.chrome.options import Options

import time
import win32api
import pandas as pd
import os



# Function to check if the file has been downloaded
def wait_for_download(directory):
    # Wait for the download to finish (check every second)
    while True:
        files = os.listdir(directory)
        if files:  # Assuming there's only one file downloading at a time
            if not any(f.endswith('.crdownload') for f in files):  # Chrome temporary download extension
                return files[0]
        time.sleep(1)

# Function to rename the downloaded file
def rename_downloaded_file(download_dir, old_name, new_name):
    old_path = os.path.join(download_dir, old_name)
    new_path = os.path.join(download_dir, new_name)
    os.rename(old_path, new_path)



# Create a WebDriver instance (in this example, using Chrome)

download_dir = "C:\\ahmedhassanali\PCRS data\shiny\Stage 1 web scraping\eligi\Raw data"  # Change this to your preferred directory


os.makedirs(download_dir, exist_ok=True)

# Chrome options for headless download
options = webdriver.ChromeOptions()
options.add_argument('--headless')
options.add_argument('--disable-gpu')  # Sometimes needed for headless mode
options.add_argument(f'--no-sandbox')  # Bypass OS security model
options.add_argument('--disable-dev-shm-usage')  # Overcome limited resource problems
options.add_experimental_option("prefs", {
    "download.default_directory": download_dir,
    "download.prompt_for_download": False,  # Disable prompts
    "download.directory_upgrade": True,
    "safebrowsing.enabled": True  # Enable to prevent download error notifications
})

driver = webdriver.Chrome(options=options)


driver.get("https://www.sspcrs.ie/portal/annual-reporting/report/eligibility")
time.sleep(5)

reportPage = driver.find_element(By.XPATH, "/html/body/div/div/section/section[2]/section/div/div[5]/div[2]/h3/form/button")
reportPage.click()
time.sleep(10)

monthList = ['202405', '202406', '202407']


for month in monthList:
    monthCombo = driver.find_element(By.XPATH, "//img[contains(@id, 'dropdownIcon')]")
    monthCombo.click()
    
    wait = WebDriverWait(driver, 10)
    monthChoice = wait.until(EC.presence_of_element_located((By.XPATH, "//*[@title='"+month+"']")))
    monthChoice.click()
    time.sleep(5)
    applyButton = driver.find_element(By.XPATH, "//*[@id='gobtn']")
    applyButton.click()
    time.sleep(5)
    
    exportLink = driver.find_element(By.XPATH, "//a[@title='Export to different format']")
    exportLink.click()
    time.sleep(5)
    
    td_element = driver.find_element(By.XPATH, "//td[text()='Excel']")
    # Click on the <td> element
    td_element.click()
    time.sleep(5)
    td_element2 = driver.find_element(By.XPATH, "//td[contains(text(), 'Excel 2007')]")
    # Click on the <td> element
    td_element2.click()
    time.sleep(5)
    # After triggering the download
    #downloaded_file_name = wait_for_download(download_dir)  # Get the name of the downloaded file
    new_file_name = month+".xlsx"  # Set your desired new file name
    rename_downloaded_file(download_dir, "MC table breakdown_dash.xlsx", new_file_name)

    print(f"File downloaded and renamed to {new_file_name}")
    time.sleep(10)
    print("Download Done")
    
driver.close()
